edit init sql

+++ sql/init.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS `order_buy` (
   `buy_ID` int(50) unsigned NOT NULL AUTO_INCREMENT,
   `trader_ID` smallint(5) unsigned NOT NULL,
   `card_ID` tinyint(3) unsigned NOT NULL,
-  `price` tinyint(3) unsigned NOT NULL,
+  `price` float unsigned NOT NULL,
   `time` datetime NOT NULL,
   PRIMARY KEY (`buy_ID`) USING BTREE
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
@@ -14,7 +14,7 @@ CREATE TABLE IF NOT EXISTS `order_sell` (
   `sell_ID` int(50) unsigned NOT NULL AUTO_INCREMENT,
   `trader_ID` smallint(5) unsigned NOT NULL,
   `card_ID` tinyint(3) unsigned NOT NULL,
-  `price` tinyint(3) unsigned NOT NULL,
+  `price` float unsigned NOT NULL,
   `time` datetime NOT NULL,
   PRIMARY KEY (`sell_ID`) USING BTREE
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
@@ -23,7 +23,7 @@ CREATE TABLE IF NOT EXISTS `trades` (
   `trade_ID` int(10) unsigned NOT NULL AUTO_INCREMENT,
   `buy_ID` int(50) unsigned NOT NULL,
   `sell_ID` int(50) unsigned NOT NULL,
-  `price` tinyint(3) unsigned NOT NULL,
+  `price` float unsigned NOT NULL,
   `card_id` tinyint(3) unsigned NOT NULL,
   `time` datetime NOT NULL,
   PRIMARY KEY (`trade_ID`) USING BTREE,
